refactor(testing): split monolithic test_app.py into focused route test files
Resolves #38

Splits tests/test_app.py into conftest.py (shared fixtures) and six
route-specific test files: test_routes_index, test_routes_screenshots,
test_routes_image, test_routes_thumb, test_routes_state, test_routes_done.
